Reduce warnings on Win64
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@30171 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/classes/tree.h
@@ -43,7 +43,7 @@ namespace Firebird {
 // This macro controls merging of nodes of all B+ trees
 // Now it merges pages only when resulting page will be 3/4 filled or less
 // Be careful while changing this expression. N=2 must always cause merge
-static inline bool NEED_MERGE(int current_count, int page_count) {
+static inline bool NEED_MERGE(size_t current_count, size_t page_count) {
 	return current_count * 4 / 3 <= page_count;
 }
 
@@ -668,7 +668,7 @@ bool BePlusTree<Value, Key, Allocator, KeyOfValue, Cmp, LeafCount, NodeCount>::a
 	// This array contains index of the element we try to add on page of each level
 	// -1 means that element is on new page
 	// In case of low memory condition we use this data to recover to innocent state
-	int recovery_map[MAX_TREE_LEVEL];
+	size_t recovery_map[MAX_TREE_LEVEL];
 			
 	if (pos == LeafCount) {
 		newLeaf->insert(0, item);
